created new feature markdown renderer
