+++ app/chat/[id]/page.tsx
@@ -9,7 +9,7 @@ import { ChatSidebar } from "@/components/chat-sidebar"
 import { NavLinks } from "@/components/nav-links"
 import { Toaster } from "@/components/UI/toaster"
 import "katex/dist/katex.min.css"
-import  Markdown from "@/components/markdown-with-latex"
+import  {Markdown} from "@/features/markdown-renderer/ui/markdown-with-latex"
 import MessageItem from "@/components/MessageItem"
 import MessageInput from "@/components/MessageInput"
 import { useAuth } from "@/features/auth/model/use-auth"
