+++ shared/const/PROGRAMMING_LANGUAGES.ts
@@ -1,0 +1,35 @@
+
+
+export const PROGRAMMING_LANGUAGES = [
+  "javascript",
+  "typescript",
+  "python",
+  "java",
+  "csharp",
+  "cpp",
+  "c",
+  "php",
+  "ruby",
+  "go",
+  "rust",
+  "swift",
+  "kotlin",
+  "scala",
+  "r",
+  "bash",
+  "sh",
+  "shell",
+  "powershell",
+  "sql",
+  "html",
+  "css",
+  "scss",
+  "less",
+  "json",
+  "yaml",
+  "xml",
+  "markdown",
+  "dockerfile",
+  "plaintext",
+  "text",
+]
