+++ shared/const/fileExtensions.ts
@@ -1,0 +1,35 @@
+
+
+export const fileExtensions: Record<string, string> = {
+      javascript: "js",
+      typescript: "ts",
+      python: "py",
+      java: "java",
+      csharp: "cs",
+      cpp: "cpp",
+      c: "c",
+      php: "php",
+      ruby: "rb",
+      go: "go",
+      rust: "rs",
+      swift: "swift",
+      kotlin: "kt",
+      scala: "scala",
+      r: "r",
+      bash: "sh",
+      sh: "sh",
+      shell: "sh",
+      powershell: "ps1",
+      sql: "sql",
+      html: "html",
+      css: "css",
+      scss: "scss",
+      less: "less",
+      json: "json",
+      yaml: "yml",
+      xml: "xml",
+      markdown: "md",
+      dockerfile: "dockerfile",
+      plaintext: "txt",
+      text: "txt",
+    }
